Moved board modules to own application
