+++ chexers/apps/board/lib/board/cell.ex
