+++ chexers/apps/board/lib/board/draw.ex
